Update members to use ResourceList instead of IndexList
