+++ ui/src/clockface/components/resource_list/ResourceList.tsx
@@ -6,8 +6,9 @@ import ResourceListHeader from 'src/clockface/components/resource_list/ResourceL
 import ResourceListSorter from 'src/clockface/components/resource_list/ResourceListSorter'
 import ResourceListBody from 'src/clockface/components/resource_list/ResourceListBody'
 import ResourceCard from 'src/clockface/components/resource_list/ResourceCard'
-import ResourceName from 'src/clockface/components/resource_list/ResourceName'
+import ResourceEditableName from 'src/clockface/components/resource_list/ResourceEditableName'
 import ResourceDescription from 'src/clockface/components/resource_list/ResourceDescription'
+import ResourceName from 'src/clockface/components/resource_list/ResourceName'
 
 interface Props {
   children: JSX.Element[] | JSX.Element
@@ -19,6 +20,7 @@ export default class ResourceList extends PureComponent<Props> {
   public static Body = ResourceListBody
   public static Card = ResourceCard
   public static Name = ResourceName
+  public static EditableName = ResourceEditableName
   public static Description = ResourceDescription
 
   public render() {
