+++ ui/src/clockface/components/resource_list/ResourceName.tsx
@@ -1,24 +1,16 @@
 // Libraries
-import React, {Component, KeyboardEvent, ChangeEvent, MouseEvent} from 'react'
-import classnames from 'classnames'
+import React, {Component, MouseEvent} from 'react'
 
 // Components
-import {Input, SpinnerContainer, TechnoSpinner} from '@influxdata/clockface'
-import {ClickOutside} from 'src/shared/components/ClickOutside'
+import {SpinnerContainer, TechnoSpinner} from '@influxdata/clockface'
 
 // Types
-import {ComponentSize} from '@influxdata/clockface'
 import {RemoteDataState} from 'src/types'
 
-// Decorators
-import {ErrorHandling} from 'src/shared/decorators/errors'
-
 interface Props {
-  onUpdate: (name: string) => void
   name: string
   onClick?: (e: MouseEvent<HTMLAnchorElement>) => void
   placeholder?: string
-  noNameString: string
   parentTestID: string
   buttonTestID: string
   inputTestID: string
@@ -26,12 +18,9 @@ interface Props {
 }
 
 interface State {
-  isEditing: boolean
-  workingName: string
   loading: RemoteDataState
 }
 
-@ErrorHandling
 class ResourceName extends Component<Props, State> {
   public static defaultProps = {
     parentTestID: 'resource-name',
@@ -44,66 +33,25 @@ class ResourceName extends Component<Props, State> {
     super(props)
 
     this.state = {
-      isEditing: false,
-      workingName: props.name,
       loading: RemoteDataState.Done,
     }
   }
 
   public render() {
-    const {
-      name,
-      hrefValue,
-      noNameString,
-      parentTestID,
-      buttonTestID,
-    } = this.props
+    const {name, hrefValue, parentTestID} = this.props
 
     return (
-      <div className={this.className} data-testid={parentTestID}>
+      <div className="resource-name" data-testid={parentTestID}>
         <SpinnerContainer
           loading={this.state.loading}
           spinnerComponent={<TechnoSpinner diameterPixels={20} />}
         >
           <a href={hrefValue} onClick={this.handleClick}>
-            <span>{name || noNameString}</span>
+            <span>{name}</span>
           </a>
         </SpinnerContainer>
-        <div
-          className="resource-name--toggle"
-          onClick={this.handleStartEditing}
-          data-testid={buttonTestID}
-        >
-          <span className="icon pencil" />
-        </div>
-        {this.input}
       </div>
     )
-  }
-
-  private get input(): JSX.Element {
-    const {placeholder, inputTestID} = this.props
-    const {workingName, isEditing, loading} = this.state
-
-    if (isEditing && loading !== RemoteDataState.Loading) {
-      return (
-        <ClickOutside onClickOutside={this.handleStopEditing}>
-          <Input
-            size={ComponentSize.ExtraSmall}
-            maxLength={90}
-            autoFocus={true}
-            spellCheck={false}
-            placeholder={placeholder}
-            onFocus={this.handleInputFocus}
-            onChange={this.handleInputChange}
-            onKeyDown={this.handleKeyDown}
-            className="resource-name--input"
-            value={workingName}
-            testID={inputTestID}
-          />
-        </ClickOutside>
-      )
-    }
   }
 
   private handleClick = (e: MouseEvent<HTMLAnchorElement>) => {
@@ -111,61 +59,6 @@ class ResourceName extends Component<Props, State> {
     if (onClick) {
       onClick(e)
     }
-  }
-
-  private handleStartEditing = (): void => {
-    this.setState({isEditing: true})
-  }
-
-  private handleStopEditing = async (): Promise<void> => {
-    const {workingName} = this.state
-    const {onUpdate} = this.props
-
-    this.setState({loading: RemoteDataState.Loading})
-    await onUpdate(workingName)
-    this.setState({loading: RemoteDataState.Done, isEditing: false})
-  }
-
-  private handleInputChange = (e: ChangeEvent<HTMLInputElement>): void => {
-    this.setState({workingName: e.target.value})
-  }
-
-  private handleKeyDown = async (
-    e: KeyboardEvent<HTMLInputElement>
-  ): Promise<void> => {
-    const {onUpdate, name} = this.props
-    const {workingName} = this.state
-
-    if (e.key === 'Enter') {
-      e.persist()
-
-      if (!workingName) {
-        this.setState({isEditing: false, workingName: name})
-
-        return
-      }
-      this.setState({loading: RemoteDataState.Loading})
-      await onUpdate(workingName)
-      this.setState({isEditing: false, loading: RemoteDataState.Done})
-    }
-
-    if (e.key === 'Escape') {
-      this.setState({isEditing: false, workingName: name})
-    }
-  }
-
-  private handleInputFocus = (e: ChangeEvent<HTMLInputElement>): void => {
-    e.currentTarget.select()
-  }
-
-  private get className(): string {
-    const {name, noNameString} = this.props
-    const {isEditing} = this.state
-
-    return classnames('resource-name', {
-      'resource-name--editing': isEditing,
-      'untitled-name': name === noNameString,
-    })
   }
 }
