+++ ui/src/clockface/styles.scss
@@ -22,6 +22,7 @@
 @import 'components/index_views/IndexList.scss';
 @import 'components/resource_list/ResourceList.scss';
 @import 'components/resource_list/ResourceDescription.scss';
+@import 'components/resource_list/ResourceEditableName.scss';
 @import 'components/resource_list/ResourceName.scss';
 @import 'components/right_click_menu/RightClick.scss';
 @import 'components/tabs/Tabs.scss';
